serial

+++ visionSerial-main/include/packet.h
@@ -11,17 +11,17 @@
 #pragma pack(2)
 struct visionMsg{
     uint16_t head;
-    float aimYaw;               // 目标Yaw
-    float aimPitch;             // 目标Pitch
+    double aimYaw;               // 目标Yaw
+    double aimPitch;             // 目标Pitch
 };
 
 struct robotMsg{
     uint16_t head;
     uint8_t mode;               // 自瞄模式 0-装甲板 1-小符 2-大符
     uint8_t foeColor;           // 敌方颜色 0-blue 1-red
-    float robotYaw;               // 自身Yaw
-    float robotPitch;             // 自身Pitch
-    float muzzleSpeed;          // 弹速
+    double robotYaw;               // 自身Yaw
+    double robotPitch;             // 自身Pitch
+    double muzzleSpeed;          // 弹速
 };
 
 union visionArray{
